Create nested route for search results component
