--- src/pages/search.js
@@ -1,31 +1,0 @@
-import Header from "../components/Header/index"
-import Footer from "../components/Footer"
-import SearchBox from "../components/shared/SearchBox"
-import {
-  ContainerMain,
-  Content,
-  PageBody,
-} from "../components/shared/Containers"
-import GenresGrid from "../components/shared/grids/GenresGrid"
-
-function Search() {
-  return (
-    <ContainerMain>
-      <Header noSearchBox />
-      <PageBody>
-        <Content flexDirection="column" alignItems="center" $marginTop="10em">
-          <SearchBox
-            placeholder="Search album, year, artist"
-            big
-            marginBottom="7em"
-          />
-          <h2>Genres</h2>
-          <GenresGrid />
-        </Content>
-      </PageBody>
-      <Footer />
-    </ContainerMain>
-  )
-}
-
-export default Search
